Created a dummy statik pkg to make the repo go-gettable
Because currently `go get` complains that it can't find the statik pkg, the command fails.
Added a dummy pkg to avoid this. It compiles fine, but fails at runtime when `go generate` isn't run. That's the downside. We might consider actually committing the real statik pkg as well (as go generate intends), but to me it feels more like a build artifact than source code.